Change type to long long to be consistent across platforms

+++ fibonacci.h
@@ -1,6 +1,6 @@
 #include <map>
 
-using fibType = long;
+using fibType = unsigned long long;
 using fibMapType = std::map<int, fibType>;
 
 fibType getFibonacci(fibMapType &fibMap, int n);
